refactor(discord): decompose recovery_engine.rs into child modules (#3834)
Pure verbatim move of the runtime-rediscovery recovery path out of the
grandfathered `recovery_engine.rs` root into a new leaf child module
`recovery_engine/runtime.rs`, mirroring the merged manual-rebind split
(PR #3957). Zero behavior change.

Moved (byte-identical except module-local `use super::*` resolution):
- `reregister_active_turn_from_inflight` (runtime recovery entry)
- `reseed_watcher_owned_finalizer_ledger` (its private helper)
- `delivered_inflight_reregister_tests` + `reregister_ledger_reseed_tests`

The entry point is re-exported from the facade root
(`pub(super) use self::runtime::reregister_active_turn_from_inflight`) so
every external path stays byte-identical: `watchers::lifecycle` via the
`recovery::` alias, the sibling `manual_rebind` call, and the root
`restore_inflight_turns` reattach call sites. Visibility preserved exactly
(`pub(in crate::services::discord)` == the prior `pub(super)`); the private
reseed helper stays private. No bare-`pub` widening.

recovery_engine.rs: 3037 -> 2935 production lines.
